seek err fix

+++ code/go/0chain.net/blobbercore/allocation/file_changer_update.go
@@ -98,6 +98,7 @@ func (nf *UpdateFileChanger) ApplyChange(ctx context.Context, change *Allocation
 	fileRef.EncryptedKey = nf.EncryptedKey
 	fileRef.ChunkSize = nf.ChunkSize
 	fileRef.IsTemp = nf.IsTemp
+	fileRef.ThumbnailFilename = nf.ThumbnailFilename
 
 	_, err = rootRef.CalculateHash(ctx, true)
 	if err != nil {
@@ -122,8 +123,14 @@ func (nf *UpdateFileChanger) CommitToFileStore(ctx context.Context) error {
 
 		if err == nil && count == 0 {
 			logging.Logger.Info("Deleting content file", zap.String("validation_root", hash))
-			if err := filestore.GetFileStore().DeleteFile(nf.AllocationID, hash, nf.Path, nf.Filename); err != nil {
-				logging.Logger.Error("FileStore_DeleteFile", zap.String("allocation_id", nf.AllocationID), zap.Error(err))
+			if hash == nf.ThumbnailHash {
+				if err := filestore.GetFileStore().DeleteFile(nf.AllocationID, hash, nf.Path, nf.ThumbnailFilename); err != nil {
+					logging.Logger.Error("FileStore_DeleteFile", zap.String("allocation_id", nf.AllocationID), zap.Error(err))
+				}
+			} else {
+				if err := filestore.GetFileStore().DeleteFile(nf.AllocationID, hash, nf.Path, nf.Filename); err != nil {
+					logging.Logger.Error("FileStore_DeleteFile", zap.String("allocation_id", nf.AllocationID), zap.Error(err))
+				}
 			}
 		}
 	}
